feat: terrain worldgen + viewport regression fix + bliss + licensing
Terrain generation (common):
- AAA deterministic worldgen modules (noise/hydrology/erosion/climate/
  materials/pipeline/preview/export) + disk chunk load + splat loader.
- Vertex-colour terrain mesh: bilinear splat sampling, altitude-resolved
  snow/water bucket (rivers read blue, not snow-white), baked ambient
  occlusion + tonal variation.

Engine viewport:
- Fix washed-out editor viewport: revert the R1 photoreal post-stack
  (Bloom/SSAO/TAA) and aggressive fog back to their committed HEAD state —
  together they flattened the scene to uniform grey.
- Keep the camera viewport-clip and the fresh-asset overlay-texture resize
  fix: the 3D fills the pane and scales correctly on window resize.

Bliss: contribution tracker + witness co-signing updates (concurrent session).

Web + licensing:
- Dual-license: PolyForm Shield + commercial (source-available).
- KYC / license / support pages, tool icons, OG image.
- Misc crate manifest updates across the workspace.

+++ eustress/crates/web/src/pages/community.rs
@@ -27,7 +27,8 @@ struct SearchUser {
     username: String,
     display_name: String,
     avatar_url: Option<String>,
-    bliss_balance: u64,
+    #[serde(default)]
+    bliss_balance: f64,
     is_verified: bool,
     follower_count: u64,
 }
